scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium==4.18.1
python-dotenv>=1.0.0
pymongo>=4.6.0
webdriver-manager==4.0.1
fake-useragent==1.4.0
schedule==1.2.1
undetected-chromedriver==3.5.5
python-anticaptcha==1.0.0
requests[socks]==2.31.0
stem==1.8.2
python-socks==2.4.4
dnspython>=2.4.0
urllib3>=2.0.0
certifi>=2023.7.22
charset-normalizer>=3.3.0
idna>=3.4
soupsieve>=2.5
typing-extensions>=4.8.0 